+++ src/main/java/eightplusone/bit/fit/global/config/AudioWebSocketConfig.java
@@ -1,7 +1,5 @@
 package eightplusone.bit.fit.global.config;
 
-import eightplusone.bit.fit.domain.streamingRoom.RoomManager;
-import eightplusone.bit.fit.global.websocket.CallHandler;
 import org.kurento.client.KurentoClient;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
@@ -9,26 +7,30 @@ import org.springframework.web.socket.config.annotation.EnableWebSocket;
 import org.springframework.web.socket.config.annotation.WebSocketConfigurer;
 import org.springframework.web.socket.config.annotation.WebSocketHandlerRegistry;
 
+import eightplusone.bit.fit.domain.streaming.RoomManager;
+import eightplusone.bit.fit.global.websocket.CallHandler;
+
 @EnableWebSocket
 @Configuration
 public class AudioWebSocketConfig implements WebSocketConfigurer {
-    @Bean
-    public RoomManager roomManager() {
-        return new RoomManager();
-    }
-    @Bean
-    public CallHandler callHandler(RoomManager roomManager) {
-        return new CallHandler(roomManager);
-    }
+	@Bean
+	public RoomManager roomManager() {
+		return new RoomManager();
+	}
 
-    @Bean
-    public KurentoClient kurentoClient()  {
-        return KurentoClient.create("ws://localhost:8888/kurento");
-    }
+	@Bean
+	public CallHandler callHandler(RoomManager roomManager) {
+		return new CallHandler(roomManager);
+	}
 
-    @Override
-    public void registerWebSocketHandlers(WebSocketHandlerRegistry registry) {
-        registry.addHandler(callHandler(roomManager()), "/call").setAllowedOrigins("*");
-    }
+	@Bean
+	public KurentoClient kurentoClient() {
+		return KurentoClient.create("ws://localhost:8888/kurento");
+	}
+
+	@Override
+	public void registerWebSocketHandlers(WebSocketHandlerRegistry registry) {
+		registry.addHandler(callHandler(roomManager()), "/call").setAllowedOrigins("*");
+	}
 
 }
